requirements.txt
Package       Version
------------- -------
autopep8      1.4.4  
pip           10.0.1 
pycodestyle   2.5.0  
python-dotenv 0.10.3 
pytz          2019.3 
setuptools    39.0.1 
